+++ workflow/mc-workflow/src/main/java/ai/mindconnect/workflow/execution/WorkflowContextFactory.java
@@ -46,7 +46,8 @@ public class WorkflowContextFactory {
                 scriptOutputWriter,
                 jsonMapper,
                 workflowDefinitionRegistry,
-                new ArrayList<>()
+                new ArrayList<>(),
+                new java.util.concurrent.ConcurrentHashMap<>()
         );
     }
 }